Refactored Meta to respective namespaces, e.g., day and weekly
