+++ src/scheduler/workers/unpublish/day/types.ts
@@ -1,0 +1,9 @@
+export interface Meta {
+  /**
+   * annual review reference for list (`List.jsonData.currentAnnualReview.reference`)
+   */
+  reference: string;
+  daysUntilUnpublish: number;
+  parsedUnpublishDate: string;
+  countryName: string;
+}
